+++ controller/scooter/scooter.go
@@ -7,7 +7,7 @@ import (
 
 // declaring the repository interface in the controller package allows us to easily swap out the actual implementation, enforcing loose coupling.
 type repository interface {
-	GetScootersNearby(latitude float64, longitude float64, distance int, limit int64) ([]model.Point, error)
+	GetScootersNearby(latitude float64, longitude float64, distance int64, limit int64) ([]model.Point, error)
 }
 
 // Controller contains the service, which contains database-related logic, as an injectable dependency, allowing us to decouple business logic from db logic.
@@ -22,7 +22,7 @@ func InitController(scooterRepo *scooterrepo.ScooterRepo) *Controller {
 	}
 }
 
-func (c *Controller) GetNearbyScooters(latitude float64, longitude float64, distance int, limit int64) ([]model.Point, error) {
+func (c *Controller) GetNearbyScooters(latitude float64, longitude float64, distance int64, limit int64) ([]model.Point, error) {
 	scooters, err := c.service.GetScootersNearby(latitude, longitude, distance, limit)
 	return scooters, err
 }
